backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
requests==2.32.5
whitenoise==6.11.0
gunicorn==23.0.0

